#13 Remove client transition handlers for synced chores.
Also merge Patches tuning state machine state transitions in one.
